refer #147 - end to end work continue: - Generating an alomost working MVC web page + controller + aspect(tiles & style is missing). Added test - Added applicationContext-test.xm for executing JUnit without web context - Validate files existence (maven/spring) when editing a project file - Other small issues

+++ openlegacy/openlegacy-providers/openlegacy-tn5250j/src/main/java/org/openlegacy/providers/tn5250j/Tn5250jTerminalEditableField.java
@@ -37,7 +37,7 @@ public class Tn5250jTerminalEditableField extends Tn5250jTerminalField {
 	@Override
 	public Class<?> getType() {
 		if (screenField.isNumeric()) {
-			return Double.class;
+			return Integer.class;
 		}
 		return String.class;
 	}
